minor changes on the todo + scoreBoard animation on dragon hits (turning score) + started implementing Dialogs -> WIP + first extra keys to implement pause
ToDo :
- make collision more accurate
- allow chopper to crash (die)
- implement scenes for next level, pauze-resume, continue after crash

+++ src/main/java/com/doubbel/javafxtest/NewBulletsLogic.java
@@ -5,7 +5,7 @@ import java.util.ArrayList;
 class NewBulletsLogic {
     private ArrayList<NewSpriteLogic> myBullets = new ArrayList<>();
 
-    NewSpriteLogic createBullet(int XPos, int YPos) {
+    NewSpriteLogic createBullet(double XPos, double YPos) {
         NewSpriteLogic newBullet = new NewBulletLogic().setSetPointTimeTaskDivider(2);
         newBullet.setLocationAbsolute(XPos, YPos);
         newBullet.setCurrentImageToIndex(8);
